+++ CHANGELOG.md
@@ -59,6 +59,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [962](https://github.com/umee-network/umee/pull/962) Streamline AccrueAllInterest
 - [967](https://github.com/umee-network/umee/pull/962) Use taylor series of e^x for more accurate interest at high APY.
 - [987](https://github.com/umee-network/umee/pull/987) Streamline x/leverage CLI tests
+- [1012](https://github.com/umee-network/umee/pull/1012) Improve negative time elapsed error message
+
+### Bug Fixes
+
+- [1018](https://github.com/umee-network/umee/pull/1018) Return nil if negative time elapsed from the last block happens.
 
 ### API Breaking